requirements.txt
discord.py==2.6.4
pyyaml==6.0.2
groq
python-dotenv
